+++ src/redturtle/prenotazioni/browser/add_edit_view.py
@@ -2,9 +2,11 @@
 from plone.dexterity.browser.add import DefaultAddForm as BaseAddForm
 from plone.dexterity.browser.add import DefaultAddView as BaseAddView
 from plone.dexterity.browser.edit import DefaultEditForm as BaseEdit
-from plone.dexterity.interfaces import IDexterityEditForm
-from plone.z3cform import layout
-from zope.interface import classImplements
+from plone.dexterity.browser.edit import DefaultEditView as BaseEditView
+
+# from Products.CMFPlone.resources import add_bundle_on_request
+from zope.interface import implementer
+from .viewlets import IHasTableOverridesMarker
 
 
 class DefaultEditForm(BaseEdit):
@@ -19,8 +21,13 @@ class DefaultEditForm(BaseEdit):
         self.widgets["week_table"].auto_append = False
 
 
-DefaultEditView = layout.wrap_form(DefaultEditForm)
-classImplements(DefaultEditView, IDexterityEditForm)
+@implementer(IHasTableOverridesMarker)
+class DefaultEditView(BaseEditView):
+    form = DefaultEditForm
+
+    # def __call__(self):
+    #     add_bundle_on_request(self.request, "week-table-overrides-widget-bundle")
+    #     return super().__call__()
 
 
 class DefaultAddForm(BaseAddForm):
@@ -35,5 +42,10 @@ class DefaultAddForm(BaseAddForm):
         self.widgets["week_table"].auto_append = False
 
 
+@implementer(IHasTableOverridesMarker)
 class DefaultAddView(BaseAddView):
     form = DefaultAddForm
+
+    # def __call__(self):
+    #     add_bundle_on_request(self.request, "week-table-overrides-widget-bundle")
+    #     return super().__call__()