+++ CHANGES.rst
@@ -38,6 +38,7 @@ Changelog
   [cekk]
 
 - Add DELETE endpoint for booking.
+- Add new field that allows to override week schedule for a certain date range.
   [cekk]
 
 1.7.1 (2023-03-28)